[snapshot]: avoid truncating size_t to int in MemStream
The read and write will handle large sizes that exceed
max int values. This cl avoids the truncating from size_t
down to int.

Bug: 417788798

+++ android/android-emu-base/android/base/files/MemStream.cpp
@@ -20,7 +20,7 @@
 namespace android {
 namespace base {
 
-MemStream::MemStream(int reserveSize) {
+MemStream::MemStream(size_t reserveSize) {
     mData.reserve(reserveSize);
 }
 
@@ -30,7 +30,7 @@ ssize_t MemStream::read(void* buffer, size_t size) {
     if (!buffer) {
         return 0;
     }
-    const auto sizeToRead = std::min<int>(size, readSize());
+    const auto sizeToRead = std::min<size_t>(size, readSize());
     memcpy(buffer, mData.data() + mReadPos, sizeToRead);
     mReadPos += sizeToRead;
     return sizeToRead;
